Implement option parsing in parser::parse_job_argument_list
This operates on the level of unexpanded tokens, and relies on a signature
being present. Since no signature for builtins and functions is present now,
it does actually nothing. Signatures for builtins and the mechanism for
declaring function with signatures will be added in subsequent commits.

Original proposal: https://github.com/fish-shell/fish-shell/issues/447